moving debug*.py to debug dir, moving some part for debugger [no ci]
