+++ inkstitch.py
@@ -8,14 +8,18 @@ import sys
 from pathlib import Path  # to work with paths as objects
 from argparse import ArgumentParser  # to parse arguments and remove --extension
 
-import toml              # to read logging configuration from toml file
-import warnings          # to control python warnings
-import logging           # to configure logging
-import logging.config    # to configure logging from dict
+if sys.version_info >= (3, 11):
+    import tomllib      # built-in in Python 3.11+
+else:
+    import tomli as tomllib
 
-import lib.debug_utils as debug_utils
-import lib.debug_logging as debug_logging
-from lib.debug_utils import safe_get    # mimic get method of dict with default value
+import logging
+
+import lib.debug.utils as debug_utils
+import lib.debug.logging as debug_logging
+from lib.debug.utils import safe_get    # mimic get method of dict with default value
+
+# --------------------------------------------------------------------------------------------
 
 SCRIPTDIR = Path(__file__).parent.absolute()
 
@@ -30,7 +34,8 @@ if old_debug_ini.exists():
 
 debug_toml = SCRIPTDIR / "DEBUG.toml"
 if debug_toml.exists():
-    ini = toml.load(SCRIPTDIR / "DEBUG.toml")  # read DEBUG.toml file if exists, otherwise use default values in ini object
+    with debug_toml.open("rb") as f:
+        ini = tomllib.load(f)  # read DEBUG.toml file if exists, otherwise use default values in ini object
 else:
     ini = {}
 # --------------------------------------------------------------------------------------------
@@ -57,52 +62,9 @@ if len(sys.argv) < 2:
         print(msg)
     exit(1)
 
-# Configure logging:
-# in release mode normally we want to ignore all warnings and logging, but we can enable it by setting environment variables
-#  - INKSTITCH_LOGLEVEL - logging level:
-#       'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'
-#  - PYTHONWARNINGS, -W - warnings action controlled by python
-#       actions: 'error', 'ignore', 'always', 'default', 'module', 'once'
-if running_as_frozen:  # in release mode
-    loglevel = os.environ.get('INKSTITCH_LOGLEVEL')  # read log level from environment variable or None
-    docpath = os.environ.get('DOCUMENT_PATH')  # read document path from environment variable (set by inkscape) or None
-
-    if docpath is not None and loglevel is not None and loglevel.upper() in ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']:
-        # end user enabled logging&warnings and redirect output to input_svg.inkstitch.log
-        logfilename = Path(docpath).with_suffix('.inkstitch.log')  # log file is created in document path
-        loglevel = loglevel.upper()
-
-        # dictConfig has access to top level variables
-        # in dict config string we use:
-        # - ext://__main__.logfilename -> access to logfilename: user.svg.inkstitch.log
-        # - ext://__main__.loglevel -> access to loglevel: user defined log level
-        # restriction: variable must be last token in string
-        logging.config.dictConfig(debug_logging.frozen_config)  # configure root logger from dict
-        logging.captureWarnings(True)                           # capture all warnings to log file with level WARNING
-    else:
-        logging.disable()                # globally disable all logging of all loggers
-        warnings.simplefilter('ignore')  # ignore all warnings
-
-else:
-    # in development mode we want to use configuration from some LOGGING.toml file
-    logging_config_file = safe_get(ini, "LOGGING", "log_config_file", default=None)
-    if logging_config_file is not None:
-        logging_config_file = Path(logging_config_file)
-        if logging_config_file.exists():
-            with open(logging_config_file, "r") as f:
-                development_config = toml.load(f) # -> dict
-        else:
-            raise FileNotFoundError(f"{logging_config_file} file not found")
-    else:                         # if LOGGING.toml file does not exist, use default logging configuration
-        loglevel = 'DEBUG'        # set log level to DEBUG
-        logfilename = SCRIPTDIR / "inkstitch.log"  # log file is created in current directory
-        development_config = debug_logging.development_config # get TOML configuration from module
-
-    # we cannot use dictConfig() as above due to restrictions in logging current implementation
-    #   - we need replace: filename = "%(SCRIPTDIR)s/inkstitch.log" -> filename = "path/inkstitch.log"
-    debug_logging.configure_logging(development_config, ini, SCRIPTDIR)  # initialize logging configuration
-    logger.info("Running in development mode")
-    logger.info(f"Using logging configuration from file: {logging_config_file}")
+# activate logging - must be done before any logging is done
+debug_logging.activate_logging(running_as_frozen, ini, SCRIPTDIR)
+# --------------------------------------------------------------------------------------------
 
 # check if running from inkscape, given by environment variable
 if os.environ.get('INKSTITCH_OFFLINE_SCRIPT', '').lower() in ['true', '1', 'yes', 'y']:
@@ -141,13 +103,19 @@ if not running_as_frozen:  # debugging/profiling only in development mode
 
 # enabling of debug depends on value of debug_type in DEBUG.toml file
 if debug_type != 'none':
-    from lib.debug import debug  # import global variable debug - don't import whole module
-    debug.enable(debug_type, SCRIPTDIR, ini)
+    from lib.debug.debugger import init_debugger
+    init_debugger(debug_type, ini)
     # check if debugger is really activated
     debug_active = bool((gettrace := getattr(sys, 'gettrace')) and gettrace())
 
+# activate logging for svg
+from lib.debug import debug  # noqa: E402  # import global variable debug - don't import whole module
+debug.enable()  # see source how enable/disable logging
+
 # log startup info
 debug_logging.startup_info(logger, SCRIPTDIR, running_as_frozen, running_from_inkscape, debug_active, debug_type, profiler_type)
+
+# --------------------------------------------------------------------------------------------
 
 # pop '--extension' from arguments and generate extension class name from extension name
 #   example:  --extension=params will instantiate Params() class from lib.extensions.
