+++ lib/debug/__init__.py
@@ -1,0 +1,11 @@
+# Authors: see git history
+#
+# Copyright (c) 2010 Authors
+# Licensed under the GNU GPL version 3.0 or later.  See the file LICENSE for details.
+
+
+# this enable:
+#    from .debug import debug
+#    from ..debug import debug
+from .debug import debug
+
